[Added] Reviews To the Data Model !
This one was a top demmand so I rushed to make it available as soon as
possible.

The reviews data will now be added to the "Apps" collections, but only
up to the last 40 reviews.

Each chunk of 40 reviews takes an extra request, so in order to make it
run faster, I will only make one extra request per app, to fetch some
reviews.